requirements.txt
backports.shutil-get-terminal-size==1.0.0
cachetools==3.1.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
decorator==4.4.1
earthengine-api==0.1.220
enum34==1.1.6
future==0.18.2
futures==3.3.0
GDAL==3.0.4
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-cloud-core==1.2.0
google-cloud-storage==1.24.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
httplib2==0.17.0
idna==2.8
ipaddress==1.0.23
ipython==5.8.0
ipython-genutils==0.2.0
lxml~=4.2.5
numpy==1.16.5
pathlib2==2.3.5
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==1.0.15
protobuf==3.11.2
psycopg2==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
Pygments==2.5.2
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dotenv==0.13.0
pytz==2019.3
requests==2.22.0
rsa==4.0
scandir==1.10.0
sentinelsat~=0.13
simplegeneric==0.8.1
simplejson==3.17.0
six==1.14.0
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.7
wcwidth==0.1.8